fix(capacity): store used-capacity as float — fixes pod-vs-client ±1 drift
The ET-CP per-pod "Projected/Actual Used" carry ×1.4 specialized weighting
(e.g. Pod 3 Mar = 109.4), but capacity_projections stored them as INTEGER and
the importer parsed them with safe_int() — so each pod was rounded and the pod
rollup drifted ±1 from the (float) per-client itemization that matches the
sheet (Mar 386 vs 386.8, Apr 394 vs 395…). The sheet has no gap because every
cell is a float.

- models.py: projected_used_capacity / actual_used_capacity Integer → Float
  (total_capacity stays int — whole editor capacities).
- migration_service: add safe_float(); parse projected/actual with it, not
  safe_int (the rounding culprit).
- run.py MART_SCHEMAS: projected/actual_used_capacity INTEGER → FLOAT (feeds
  editorial_int_capacity_pod_months + the phase-1 flat mart; the raw BQ mirror
  type comes from the model via schema_for_model).
- main.py: idempotent Neon migration ALTER … TYPE double precision.
- schemas.py / routers/capacity.py: response types int → float.

Distinct from the non-client-rows gap (Feb/Jul); this cures the ±1 in every
month. Takes effect after deploy + an ET CP capacity sync re-lands unrounded
values + a warehouse publish.

+++ backend/app/models.py
@@ -134,8 +134,12 @@ class CapacityProjection(Base):
     year: Mapped[int] = mapped_column(Integer, nullable=False)
     month: Mapped[int] = mapped_column(Integer, nullable=False)
     total_capacity: Mapped[int | None] = mapped_column(Integer)
-    projected_used_capacity: Mapped[int | None] = mapped_column(Integer)
-    actual_used_capacity: Mapped[int | None] = mapped_column(Integer)
+    # Float, not Integer: the sheet's per-pod "Projected/Actual Used" carry ×1.4
+    # specialized weighting (e.g. Pod 3 = 109.4). Rounding each pod to int and
+    # summing lost the fractions, so the pod rollup drifted ±1 from the (float)
+    # per-client itemization that matches the sheet. See tasks/todo.md §4.
+    projected_used_capacity: Mapped[float | None] = mapped_column(Float)
+    actual_used_capacity: Mapped[float | None] = mapped_column(Float)
     version: Mapped[str | None] = mapped_column(String(50))
     notes: Mapped[str | None] = mapped_column(Text)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
